requirements.txt
chainlit==2.5.5
openai==1.84.0
python-dotenv==1.1.0